Subindo Dockerfile

+++ database/migrations/2024_06_29_005742_create_produtos_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('dataProduto');
             $table->integer('quantidade');
             $table->float('precoUnitario');
-            $table->integer('id_fornecedor');
+            $table->bigIncrements('id_fornecedor');
             $table->foreign('id_fornecedor')->references('id')->on('fornecedors')->onDelete('cascade')->onUpdate('cascade');
             $table->timestamps();
         });
